+++ src/gd/twohundred/jvb/components/interfaces/Emulable.java
@@ -1,5 +1,5 @@
 package gd.twohundred.jvb.components.interfaces;
 
 public interface Emulable extends Resetable {
-    int tick(int targetCycles);
+    long tick(long targetCycles);
 }
